// lib/api/generated/models/utils-entity-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const UtilsEntityType = {
    EntityTypeBrand: 'brand',
    EntityTypeCampaign: 'campaign',
    EntityTypeCreator: 'creator',
    EntityTypeAdminCampaignApproval: 'admin_campaign_approval',
    EntityTypeBrandCampaignDecision: 'brand_campaign_decision',
    EntityTypeVideoSubmissionStatusUpdate: 'video_submission_status_update',
    EntityTypeBrandVideoDecision: 'brand_video_decision'
} as const;

export type UtilsEntityType = typeof UtilsEntityType[keyof typeof UtilsEntityType];